requirements.txt
selenium 4.25.0
python 3.13.3
pytest-8.4.1
allure-pytest-2.15.0
openpyxl 3.1.5

